컨트랙트 배포, 추첨, 당첨 구현

+++ src/entry/entities/entry.entity.ts
@@ -19,7 +19,7 @@ export class Entry {
     @Column({ type: 'bigint', nullable: false, unique: false })
     user_id: number;
 
-    @Column({ type: 'float', nullable: false, unique: false, default: 100 })
+    @Column({ type: 'int', nullable: false, unique: false, default: 100 })
     raffle_index: number
 
     @CreateDateColumn()
